\name{ccmm_test_data}
\alias{ccmm_test_data}
\docType{data}
\title{
Test Data
}
\description{
Contains artificial 200 samples with a continuous outcome variable y, a continuous treatment tr, 20 compositional mediators M and 2 covariates X. The true direct and indirect effects of treatment on the outcome both are 1.00. The true component-wise indirect effects (M1-M20) are 0.693, -0.425, 0.135, -0.057, -0.268, 0.970, -0.843, 0.805, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000.
}
\usage{data(ccmm_test_data)}
\keyword{datasets}
